haku-console: regroup tool rendering server-first (tool_rendering/<server>/{requests,responses})
Pure move/rename: replace the direction-first tool_previews/ + tool_results/
trees with one server-first tool_rendering/ tree — each server directory owns
its widgets (requests.tsx for arguments, responses.tsx for results;
request-only servers just have no responses module), with the shared
variant.tsx / entry.tsx / result_entry.tsx leaves and one index.tsx holding
both registries (toolPreview / toolActionDescription / toolResultPreview).
Server-id constants live once per server, in requests.tsx. Tests move
alongside their modules; the two registry specs merge into one index.test.ts.
BUILD graph, tailwind/tsc/vitest aggregates, consumers, and docs updated to
the new labels and paths. No behavior changes.
