+++ haku/console/frontend/tool_action_line.tsx
@@ -1,6 +1,6 @@
 import { Text } from "@mantine/core";
 
-import { toolActionDescription } from "./tool_previews";
+import { toolActionDescription } from "./tool_rendering/index.tsx";
 
 /** The card's identity line. A registered tool renders its own action description ("Gmail: Draft
  * email", "kubectl: Delete Pod" — destructive ones in red); a tool with no widget falls back to
