+++ haku/console/frontend/tool_rendering/google_calendar/requests.tsx
@@ -8,12 +8,12 @@ import { Anchor, Loader, Stack, Text } from "@mantine/core";
 import { useEffect, useState } from "react";
 import type { z } from "zod";
 
-import { fetchCalendarSummary, type CalendarSummary } from "../calendar_client.ts";
-import { Field } from "../field.tsx";
-import { BellIcon, CalendarIcon, ClockIcon, MapPinIcon, UsersIcon } from "../icons.tsx";
-import { mcpToolSchema } from "../mcp_tool_schema.ts";
-import { definePreview, type ToolPreview } from "./entry.tsx";
-import type { PreviewProps } from "./variant.tsx";
+import { fetchCalendarSummary, type CalendarSummary } from "../../calendar_client.ts";
+import { Field } from "../../field.tsx";
+import { BellIcon, CalendarIcon, ClockIcon, MapPinIcon, UsersIcon } from "../../icons.tsx";
+import { mcpToolSchema } from "../../mcp_tool_schema.ts";
+import { definePreview, type ToolPreview } from "../entry.tsx";
+import type { PreviewProps } from "../variant.tsx";
 
 export const GOOGLE_CALENDAR_SERVER_ID = "google_calendar";
