+++ haku/console/frontend/tool_rendering/haku_routine/requests.tsx
@@ -7,10 +7,10 @@
 import { Text } from "@mantine/core";
 import { z } from "zod";
 
-import { Field } from "../field.tsx";
-import { mcpToolSchema } from "../mcp_tool_schema.ts";
-import { definePreview, type ToolPreview } from "./entry.tsx";
-import { clampBlock, type PreviewProps } from "./variant.tsx";
+import { Field } from "../../field.tsx";
+import { mcpToolSchema } from "../../mcp_tool_schema.ts";
+import { definePreview, type ToolPreview } from "../entry.tsx";
+import { clampBlock, type PreviewProps } from "../variant.tsx";
 
 export const HAKU_ROUTINE_SERVER_ID = "haku_routine";
