+++ haku/console/frontend/tool_rendering/result_entry.tsx
@@ -1,12 +1,12 @@
 // A registry entry for one tool's *result*: the zod schema for the tool's unwrapped result
 // payload paired with a renderer over the *parsed*, typed value — the result-side mirror of
-// tool_previews/entry.tsx. The registry (index.tsx) safeParses once per dispatch and hands the
+// entry.tsx. The registry (index.tsx) safeParses once per dispatch and hands the
 // widget already-validated data, so no widget repeats `schema.safeParse(payload)`. Leaf module
 // (no widget deps) so index.tsx and every per-server module import it without a cycle.
 import type { ReactNode } from "react";
 import { z } from "zod";
 
-import type { PreviewVariant } from "../tool_previews/variant.tsx";
+import type { PreviewVariant } from "./variant.tsx";
 
 export type ToolResultPreview<S extends z.ZodTypeAny = z.ZodTypeAny> = {
   schema: S;
