--- haku/console/frontend/tool_results/BUILD.bazel
@@ -1,121 +1,0 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-
-package(default_visibility = ["//visibility:public"])
-
-# Per-server tool-call *result* widgets — the result-side mirror of ../tool_previews. Each
-# module owns one server's widgets and exports a { toolName -> {schema, render} } map;
-# `tool_results` (index.tsx) dispatches to them via a serverId registry, safeParsing the
-# unwrapped result payload once and handing each widget already-typed data.
-
-# A registry entry (the tool's zod result schema + a renderer over its parsed value) plus the
-# defineResultPreview/renderResultPreview helpers and the CallToolResult unwrap. Leaf, so
-# index.tsx and every widget import it without a cycle.
-js_library(
-    name = "entry",
-    srcs = ["entry.tsx"],
-    deps = [
-        "//:node_modules/react",
-        "//:node_modules/zod",
-        "//haku/console/frontend/tool_previews:variant",
-    ],
-)
-
-js_library(
-    name = "gmail",
-    srcs = ["gmail.tsx"],
-    deps = [
-        ":entry",
-        "//:node_modules/@mantine/core",
-        "//:node_modules/react",
-        "//:node_modules/zod",
-    ],
-)
-
-js_library(
-    name = "google_calendar",
-    srcs = ["google_calendar.tsx"],
-    deps = [
-        ":entry",
-        "//:node_modules/@mantine/core",
-        "//:node_modules/react",
-        "//:node_modules/zod",
-    ],
-)
-
-js_library(
-    name = "grocy",
-    srcs = ["grocy.tsx"],
-    deps = [
-        ":entry",
-        "//:node_modules/@mantine/core",
-        "//:node_modules/react",
-        "//:node_modules/zod",
-        "//haku/console/frontend/tool_previews:variant",
-    ],
-)
-
-js_library(
-    name = "tool_results",
-    srcs = ["index.tsx"],
-    deps = [
-        ":entry",
-        ":gmail",
-        ":google_calendar",
-        ":grocy",
-        "//:node_modules/react",
-        "//haku/console/frontend/tool_previews:gmail",
-        "//haku/console/frontend/tool_previews:google_calendar",
-        "//haku/console/frontend/tool_previews:grocy",
-        "//haku/console/frontend/tool_previews:variant",
-    ],
-)
-
-# Sources Tailwind must scan for class names; the parent's tailwind_content_srcs
-# concatenates this into the one file the stylesheet's @source lists.
-filegroup(
-    name = "tailwind_srcs",
-    srcs = glob(["*.tsx"]),
-)
-
-# One vitest spec per module (each test parallels the module it tests), wrapped as its own
-# js_library so the dep graph stays per-file; the parent's aggregate tsc_test + bridge_test
-# consume them (a js_test can't take a source file from another package directly, only via a
-# library).
-js_library(
-    name = "entry_test",
-    srcs = ["entry.test.ts"],
-    deps = [":entry"],
-)
-
-js_library(
-    name = "gmail_test",
-    srcs = ["gmail.test.ts"],
-    deps = [
-        ":entry",
-        ":gmail",
-    ],
-)
-
-js_library(
-    name = "google_calendar_test",
-    srcs = ["google_calendar.test.ts"],
-    deps = [
-        ":entry",
-        ":google_calendar",
-    ],
-)
-
-js_library(
-    name = "grocy_test",
-    srcs = ["grocy.test.ts"],
-    deps = [
-        ":entry",
-        ":grocy",
-    ],
-)
-
-js_library(
-    name = "index_test",
-    srcs = ["index.test.ts"],
-    deps = [":tool_results"],
-)
